+++ R/incidence.from.adjacency.R
@@ -5,15 +5,16 @@
 #'    a given generative model
 #'
 #' @param G A symmetric, binary adjacency matrix of class `matrix` or `Matrix`,
-#'    a `data.frame` containing a symbolic edge list in the first two columns,
 #'    or an undirected, unweighted unipartite graph of class {\link{igraph}}.
 #' @param k integer: Number of artifacts to generate
 #' @param p numeric: Tuning parameter for artifacts, 0 <= p <= 1
-#' @param d numeric: Number of dimensions in Blau space, d >= 2
+#' @param maximal boolean: Should teams/groups models be seeded with *maximal* cliques?
+#' @param blau.param vector: Vector of parameters that control blau space (see details)
 #' @param model string: Generative model, one of c("team", "group", "blau") (see details)
-#' @param class string: Return object as `matrix`, `igraph`, or `edgelist`. If `NULL`, object is returned in the same class as `G`.
+#' @param class string: Return object as `matrix`, `Matrix`, or `igraph`. If `NULL`, object is returned in the same class as `G`.
+#' @param narrative boolean: TRUE if suggested text & citations should be displayed.
 #'
-#' @return An incidence matrix of class `matrix`, or a bipartite graph as an edgelist of {\link{igraph}} object.
+#' @return An incidence matrix of class `matrix` or `Matrix`, or a bipartite graph of class {\link{igraph}}.
 #'
 #' @details
 #' Given a unipartite network composed of *i agents* (i.e. nodes) that can be represented by an *i x i* adjacency
@@ -30,15 +31,21 @@
 #'
 #' The **Blau Space Model** (`model == "blau"`) mirrors an organization (the artifact) recruiting members from social
 #'     space, where those within the organization's niche join with probability `p`, and those outside the niche join
-#'     with probability 1-`p`.
+#'     with probability 1-`p`. `blau.param` is a vector containing three values that control the characteristics of the
+#'     blau space. The first value is the space's dimensionality. The second two values are shape parameters of a Beta
+#'     distribution that describes niche sizes. The default is a two-dimensional blau space, with organization niche
+#'     sizes that are strongly positively skewed (i.e., many specialist organizations, few generalists).
+#'
+#' @references {Neal, Z. P. 2022. The Duality of Networks and Foci: Generative Models of Two-Mode Networks from One-Mode Networks. *arXiv:2204.13670* \[cs.SI\]. \doi{10.48550/arXiv.2204.13670}}
+#' @references {Neal, Z. P. 2022. incidentally: An R package to generate incidence matrices and bipartite graphs. *OSF Preprints* \doi{10.31219/osf.io/ectms}}
 #'
 #' @export
 #'
 #' @examples
 #' G <- igraph::erdos.renyi.game(10, .4)
 #' I <- incidence.from.adjacency(G, k = 1000, p = .95,
-#'                               model = "team")
-incidence.from.adjacency <- function(G, k = 1, p = 1, d = 2, model = "team", class = NULL) {
+#'                               model = "team", narrative = TRUE)
+incidence.from.adjacency <- function(G, k = 1, p = 1, blau.param = c(2,1,10), maximal = TRUE, model = "team", class = NULL, narrative = TRUE) {
 
   #### Sampling function, to allow sampling from a vector with one entry ####
   sample.vec <- function(x, ...) x[sample(length(x), ...)]
@@ -46,15 +53,14 @@ incidence.from.adjacency <- function(G, k = 1, p = 1, d = 2, model = "team", cla
   #### Parameter checks ####
   if (is.null(class) & methods::is(G, "igraph")) {class <- "igraph"}
   if (is.null(class) & methods::is(G, "matrix")) {class <- "matrix"}
-  if (is.null(class) & methods::is(G, "Matrix")) {class <- "matrix"}
-  if (is.null(class) & methods::is(G, "data.frame")) {class <- "edgelist"}
+  if (is.null(class) & methods::is(G, "Matrix")) {class <- "Matrix"}
   if (!is.numeric(k)) {stop("k must be numeric")}
-  if (!is.numeric(d)) {stop("d must be numeric")}
   if (!is.numeric(p)) {stop("p must be numeric")}
-  if (d%%1!=0 | d < 2) {stop("d must be an integer greater than 1")}
   if (p < 0 | p > 1) {stop("p must be between 0 and 1")}
   if (!(model %in% c("team", "group", "blau"))) {stop("model must be one if c(\"team\", \"group\", \"blau\")")}
-  if (!(class %in% c("matrix", "igraph"))) {stop("model must be one if c(\"matrix\", \"igraph\"")}
+  if (!(class %in% c("matrix", "Matrix", "igraph"))) {stop("class must be one if c(\"matrix\", \"Matrix\", \"igraph\")")}
+  if (blau.param[1]%%1!=0 | blau.param[1]<2) {stop("The first blau.param must be an integer greater than 1")}
+  if (blau.param[2]<0 | blau.param[3]<0) {stop("The second and third blau.param must be positive")}
 
   #### Check input, get names, convert to igraph ####
   if (methods::is(G, "igraph")) {
@@ -76,22 +82,13 @@ incidence.from.adjacency <- function(G, k = 1, p = 1, d = 2, model = "team", cla
     if (!is.null(rownames(G))) {nodes <- rownames(G)} else {nodes <- 1:nrow(G)}
     G <- igraph::graph_from_adjacency_matrix(G,mode="undirected")
   }
-  if (methods::is(G, "data.frame")) {
-    if (ncol(G)!=2) {stop("the edgelist must have two columns")}
-    G <- igraph::graph_from_data_frame(G, directed = F)
-    if (!is.null(igraph::V(G)$name)) {nodes <- igraph::V(G)$name} else {nodes <- 1:(igraph::gorder(G))}
-  }
 
   I <- as.matrix(1:(igraph::gorder(G)))  #Create empty incidence with numeric row labels
 
   #### Team model (Guimera et al., 2005) ####
   if (model == "team") {
 
-    cliques <- igraph::cliques(G, min=2)  #List of all cliques
-    #This step will be slow for large/dense graphs. For these graphs, consider using this approximation
-    # - List all maximal cliques
-    # - In loop, sample one maximal clique
-    # - In loop, sample between 2 and N nodes from the maximal clique
+    if (maximal) {cliques <- igraph::max_cliques(G, min=2)} else {cliques <- igraph::cliques(G, min=2)}  #List all (maximal) cliques
 
     for (i in 1:k) {                              #For each new team k:
       clique <- sample(1:length(cliques),1)       #Pick a prior team
@@ -130,11 +127,7 @@ incidence.from.adjacency <- function(G, k = 1, p = 1, d = 2, model = "team", cla
         return(candidates)
       }
 
-      cliques <- igraph::cliques(G, min=2)  #List of all cliques
-      #This step will be slow for large/dense graphs. For these graphs, consider using this approximation
-      # - List all maximal cliques
-      # - In loop, sample one maximal clique
-      # - In loop, sample between 2 and N nodes from the maximal clique
+      if (maximal) {cliques <- igraph::max_cliques(G, min=2)} else {cliques <- igraph::cliques(G, min=2)}  #List all (maximal) cliques
 
       for (i in 1:k) {                                                  #For each new group k:
         members <- as.numeric(cliques[[sample(1:length(cliques),1)]])   #Sample a clique (initial members)
@@ -157,20 +150,25 @@ incidence.from.adjacency <- function(G, k = 1, p = 1, d = 2, model = "team", cla
   if (model == "blau") {
 
     if (!igraph::is.connected(G)) {stop("the blau space model requires that the network be connected")}
-    coords <- igraph::layout_with_mds(G, dim = d)  #Get coordinates in Blau Space
+    coords <- igraph::layout_with_mds(G, dim = blau.param[1])  #Get coordinates in Blau Space
     D <- as.matrix(stats::dist(coords))            #Compute distances between nodes in Blau Space
-    diag(D) <- NA
 
     i <- 1
-    while (dim(I)[2] < (k + 1)) {                    #Until `k` organizations are created, for each organization i:
-      leader <- sample(1:igraph::gorder(G),1)        #Pick node to serve as organization leader and niche center
-      R <- (stats::rbeta(1,2,5) * (max(D,na.rm=T) - min(D,na.rm=T))) + min(D,na.rm=T)  #Pick niche radius
-      prob <- ifelse(D[leader,] <= R, p, 1-p)        #Probability of joining, depending on whether inside or outside niche
-      prob[is.na(prob)] <- 1                         #Leader always joins
-      members <- stats::rbinom(length(prob),1,prob)  #Nodes join organization with given probability
-      if (sum(members) > 1) {      #If more than one person joins organization i
-        I <- cbind(I, members)     #Add this organization's member list to I
-        i <- i + 1}                #Go to the next organization
+    while (dim(I)[2] <= k) {                         #Until `k` organizations are created, for each organization i:
+      R <- (stats::rbeta(1,blau.param[2],blau.param[3]) * (max(D,na.rm=T) - min(D,na.rm=T))) + min(D,na.rm=T)  #Pick niche radius
+      center <- sample(1:igraph::gorder(G),1)        #Pick a node to serve as niche center
+      dat <- data.frame(dist = D[center,])           #Start dataframe, add nodes' distance from niche center
+      dat$inside <- dat$dist <= R                    #...add nodes' location inside or outside niche
+      dat$member <- NA                               #...add initial memberships in organization
+      dat$member[which(dat$inside)] <- stats::rbinom(sum(dat$inside),1,p)            #Each node inside niche joins with probability p
+      while (sum(dat$member,na.rm=T)<sum(dat$inside) & sum(is.na(dat$member))!=0) {  #While spaces & recruits are available...
+        recruit <- min(dat$dist[which(is.na(dat$member))])                           #...find recruit nearest to niche
+        dat$member[which(dat$dist==recruit)] <- stats::rbinom(1,1,1-p)               #...attempt to recruit
+      }
+      dat$member[which(is.na(dat$member))] <- 0  #Non-recruits are non-members
+      if (sum(dat$member) > 1) {   #If more than one person joins organization i...
+        I <- cbind(I, dat$member)  #...add this organization's member list to I
+        i <- i + 1}                #...go to the next organization
     }
   }
 
@@ -181,7 +179,24 @@ incidence.from.adjacency <- function(G, k = 1, p = 1, d = 2, model = "team", cla
     colnames(I) <- c(paste0("k", 1:ncol(I)))  #Insert column names
   }
   if (class == "igraph") {I <- igraph::graph_from_incidence_matrix(I)}
-  if (class == "edgelist") {I <- data.frame(igraph::as_edgelist(igraph::graph_from_incidence_matrix(I)))}
+  if (class == "Matrix"){I <- Matrix::Matrix(I)}
+
+  #Display narrative if requested
+  if (narrative) {
+    version <- utils::packageVersion("incidentally")
+    type <- " newly-formed teams"
+    if (model == "group") {type <- " newly-formed groups"}
+    if (model == "blau") {type <- " newly-formed organizations"}
+    if (class == "igraph") {text <- paste0("We used the incidentally package for R (v", version, "; Neal, 2022a) to generate a random bipartite graph from a unipartite graph. The bipartite graph represents the memberships of the ", igraph::gorder(G), " nodes from the unipartite network in ", k, type, " (Neal, 2022b).")}
+    if (class != "igraph") {text <- paste0("We used the incidentally package for R (v", version, "; Neal, 2022a) to generate a random incidence matrix from a unipartite graph. The incidence matrix represents the memberships of the ", igraph::gorder(G), " nodes from the unipartite network (rows) in ", k, type, " (columns; Neal, 2022b).")}
+    message("")
+    message("=== Suggested manuscript text and citations ===")
+    message(text)
+    message("")
+    message("Neal, Z. P. (2022a). incidentally: An R package to generate incidence matrices and bipartite graphs. OSF Preprints. https://doi.org/10.31219/osf.io/ectms")
+    message("Neal, Z. P. (2022b). The Duality of Networks and Foci: Generative Models of Two-Mode Networks from One-Mode Networks. arXiv:2204.13670 [cs.SI]. https://doi.org/10.48550/arXiv.2204.13670")
+  }
+
   return(I)  #Return the bipartite graph with row labels
 }