+++ NEWS.md
@@ -2,6 +2,11 @@
 title: "NEWS"
 output: html_document
 ---
+## incidentally 1.0.0
+
+* all functions revised, all documentation updated
+* added `incidence.from.congress()` to generate incidence matrices of bill sponsorship in the US Congress
+
 ## incidentally 0.9.0
 
 * initial release